# Conflicts:
#	Assets/TowerDefense/Scripts/TowerAgent/GunsManager.cs

+++ Assets/TowerDefense/Scripts/Bullets/Bullet.cs
@@ -7,31 +7,59 @@ public class Bullet : MonoBehaviour
     GameObject target;
     float speed;
     float damage;
+    TowerType type;
+    bool readyToDestroy;
 
-    public void setBulletInfo(GameObject t, float s, float d)
+    public Bullet(GameObject t, float s, float d, TowerType typ)
+    {
+        setBulletInfo(t, s, d, typ);
+    }
+
+    public void setBulletInfo(GameObject t, float s, float d, TowerType typ)
     {
         target = t;
         speed = s;
         damage = d;
+        type = typ;
+        readyToDestroy = false;
     }
 
     private void Update()
     {
         float step = speed * Time.deltaTime;
         transform.position = Vector3.MoveTowards(transform.position, target.transform.position, step);
+        transform.LookAt(target.transform);
+        if (readyToDestroy)
+            destroyBullet();
     }
 
     private void OnCollisionEnter(Collision collision)
     {
         if (collision.gameObject.tag == "Enemy")
         {
-            collision.gameObject.BroadcastMessage("ApplyDamage", damage);
+            collision.gameObject.BroadcastMessage("ApplyDamage", this);
         }
-        destroyBullet();
+        else if (collision.gameObject.tag != "Bullet")
+            readyToDestroy = true;
     }
 
     private void destroyBullet()
     {
         Destroy(this.gameObject);
+    }
+
+    public void SetReadyToDestroy()
+    {
+        readyToDestroy = true;
+    }
+        
+    public float GetDamage()
+    {
+        return damage;
+    }
+
+    public TowerType GetBulletType()
+    {
+        return type;
     }
 }
